+++ packages/install-expo-modules/src/plugins/android/__tests__/fixtures/MainApplication-rn083-updated.kt
@@ -1,4 +1,4 @@
-package com.myapp
+package com.helloworld
 import android.content.res.Configuration
 import expo.modules.ApplicationLifecycleDispatcher
 import expo.modules.ExpoReactHostFactory