+++ packages/install-expo-modules/src/plugins/android/__tests__/fixtures/MainApplication-rn083.kt
@@ -1,4 +1,4 @@
-package com.myapp
+package com.helloworld
 
 import android.app.Application
 import com.facebook.react.PackageList